refactor(input): 입력 추상 클래스 패키지 변경
- 입력 추상 클래스를 input 패키지에 포함하도록 변경
